% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_.R
\name{scale_fill_sphsu}
\alias{scale_fill_sphsu}
\title{SPHSU fill scales}
\usage{
scale_fill_sphsu(palette = "mixed", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{Character. Calls a palette from \link{sphsu_cols}. Current schemes are "main", "cool", "hot" and "mixed".}

\item{discrete}{Logical. Toggle to represent discrete or continuous data.}

\item{reverse}{Logical. Reverse order of colours at assignment.}

\item{...}{other arguments passed on to \link[ggplot2]{discrete_scale} (for discrete data) or \link[ggplot2]{scale_colour_gradient} (for continuous).}
}
\description{
Unit colours for \link[ggplot2]{ggplot2} graphs
}
\details{
These functions call colour palettes based on Unit colourschemes
}
\examples{

ggplot(iris, aes(Sepal.Width, Sepal.Length, col = Species)) +
  geom_point(size = 4) +
  scale_colour_sphsu()

ggplot(iris, aes(Sepal.Width, Sepal.Length, col = Sepal.Length)) +
  geom_point(size = 4) +
  scale_colour_sphsu("hot", discrete = FALSE)

ggplot(iris, aes(Sepal.Width, Sepal.Length, col = Sepal.Length)) +
  geom_point(size = 4) +
  scale_colour_sphsu("cool", discrete = FALSE)

ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
  geom_bar() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_fill_sphsu(palette = "mixed", guide = "none")


}
