% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphsu_.R
\name{sphsu_cols}
\alias{sphsu_cols}
\title{Select SPHSU colours}
\usage{
sphsu_cols(..., subset = "all", names = TRUE)
}
\arguments{
\item{...}{Character. Colour names separated by commas. Left blank this will return all colours in subset.}

\item{subset}{Character. "all", "main" or "more" can be called (this only needs used in place of colour names above).}

\item{names}{Logical. If passing directly to \link[ggplot2]{scale_manual} then set this as \code{TRUE} to remove names and allow assignment to plotted variables.}
}
\description{
This function allows you to select a colour or a vector of colours by name, or to specify a 'main' or 'more' subset.
}
\details{
The names of the standard Social and Public Health Sciences Unit colours can be passed as arguments to return a named vector of corresponding hex strings.
}
\examples{
sphsu_cols()
sphsu_cols("University Blue", "Leaf", "Moss")
sphsu_cols(subset = "main")
}
\author{
Andy Baxter a.baxter.1@research.gla.ac.uk
}
