% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphsu_.R
\name{sphsu_pal}
\alias{sphsu_pal}
\title{SPHSU palettes (for passing to other functions)}
\usage{
sphsu_pal(palette = "mixed", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of pallete to use. `"main"`, `"mixed"`, `"planet"` or `"all"`}

\item{reverse}{Reverse order. defaults to `FALSE`}

\item{...}{Additional arguments passed to colorRampPalette}
}
\description{
SPHSU palettes (for passing to other functions)
}
