% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sphsu_palettes}
\alias{sphsu_palettes}
\title{Palettes list}
\format{A list of 5 named vactors, organised by pallete}
\usage{
sphsu_palettes
}
\description{
Arranged into four palette: "Main", "Cool", "Hot", "Mixed" and "Planet"
}
\keyword{datasets}
