% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphsu_show_colours.R
\name{sphsu_show_colours}
\alias{sphsu_show_colours}
\title{See colours and colour names}
\usage{
sphsu_show_colours()
}
\value{
A pie chart ggplot graph object
}
\description{
Displays all colours in a pie chart
}
\examples{
sphsu_show_colours()
}
