% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_.R
\name{theme_sphsu_light}
\alias{theme_sphsu_light}
\title{SPHSU light theme}
\usage{
theme_sphsu_light()
}
\description{
Based on theme_light
}
\examples{
ggplot(iris, aes(Sepal.Width, Sepal.Length, col = Sepal.Length)) +
geom_point(size = 4) +
scale_colour_sphsu("hot", discrete = FALSE) +
theme_sphsu_light() +
facet_wrap(~Species)


}
