% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphsu_.R
\name{sphsu_pal}
\alias{sphsu_pal}
\title{SPHSU palettes (for passing to other functions)}
\usage{
sphsu_pal(palette = "mixed", reverse = FALSE, ...)
}
\description{
SPHSU palettes (for passing to other functions)
}
