% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{sphsu_palettes}
\alias{sphsu_palettes}
\title{Base set of 5 palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
sphsu_palettes
}
\description{
A list of five palettes for use in creating graphs
}
\keyword{datasets}
