% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPHSUgraphs.R
\docType{package}
\name{SPHSUgraphs}
\alias{SPHSUgraphs}
\alias{SPHSUgraphs-package}
\title{Colour and fill palettes for Social and Public Health Sciences Unit}
\description{
Generate graph aesthetics for ggplot2
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewbaxter439/SPHSUgraphs}
  \item Report bugs at \url{https://github.com/andrewbaxter439/SPHSUgraphs/issues}
}

}
\author{
\strong{Maintainer}: Andy Baxter \email{a.baxter.1@reserach.gla.ac.uk}

}
